src/requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-tables2==2.0.5
django-user-agents==0.4.0
gunicorn==19.9.0
olefile==0.46
pytz==2018.7
whitenoise==4.1.3
gTTS==2.0.4
gTTS-token==1.1.3
SpeechRecognition==3.8.1
google-cloud-texttospeech
google-cloud-translate
django-autotranslate
django-modeltranslation==0.13.3
